+++ src/components/nav-bar.svelte
@@ -1,0 +1,9 @@
+<nav class="menu">
+	<a href="/" class="menu-index">
+		<h1>SuperRentals</h1>
+	</a>
+	<div class="links">
+		<a href="/about" class="menu-about"> About </a>
+		<a href="/getting-in-touch" class="menu-contact"> Contact </a>
+	</div>
+</nav>