+++ src/routes/+layout.svelte
@@ -1,0 +1,6 @@
+<script>
+	import NavBar from '@components/nav-bar.svelte';
+</script>
+
+<NavBar />
+<slot />