+++ src/routes/+page.svelte
@@ -1,6 +1,9 @@
-<div class="jumbo">
-	<div class="right tomster" />
+<script>
+  import Jumbo from '@components/jumbo.svelte';
+</script>
+
+<Jumbo>
 	<h2>Welcome to Super Rentals!</h2>
 	<p>We hope you find exactly what you're looking for in a place to stay.</p>
   <a href="/about" class="button">About Us</a>
-</div>
+</Jumbo>